refactor: Create libraries and better hierarchy
