+++ prj.conf
@@ -2,7 +2,7 @@ CONFIG_BT=y
 CONFIG_BT_PERIPHERAL=y
 CONFIG_BT_SMP=y
 CONFIG_TINYCRYPT=y
-CONFIG_BT_DEVICE_NAME="HouseEnvSensBLE"
+CONFIG_BT_DEVICE_NAME="House Environmental Monitor"
 CONFIG_BT_DIS=y
 CONFIG_BT_DIS_PNP=n
 CONFIG_BT_BAS=y
